refactor(promql): split the engine module by evaluation concern
Pure code motion of engine/mod.rs (3256 lines) into focused files, each
carrying its own tests:

- mod.rs: the Engine state, entry points, and exec_expr dispatch
- columns.rs: the label-column pre-scan of the query AST
- selector.rs: selector evaluation, data loading, and result assembly
- aggregate.rs: aggregation dispatch over the fused evaluators
- call.rs: function-call dispatch and argument helpers
- test_util.rs: fixtures shared by the module's tests

Each file's module doc states which Engine fields it touches; methods
called across sibling files became pub(super), everything else keeps
its visibility and crate::engine::Engine stays unchanged.
