+++ src/promql/src/engine/call.rs
@@ -1,0 +1,710 @@
+// Copyright 2026 OpenObserve Inc.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+//
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+//! PromQL function-call dispatch and argument helpers. Touches only
+//! `eval_ctx` besides recursing through `exec_expr`.
+
+use std::str::FromStr;
+
+use config::meta::promql::value::*;
+use datafusion::error::{DataFusionError, Result};
+use hashbrown::HashSet;
+use promql_parser::parser::{Expr as PromExpr, Function, FunctionArgs};
+
+use super::Engine;
+use crate::functions;
+
+impl Engine {
+    async fn call_expr_first_arg(&mut self, args: &FunctionArgs) -> Result<Value> {
+        self.exec_expr(args.args.first().expect("Missing arg 0"))
+            .await
+    }
+
+    async fn call_expr_second_arg(&mut self, args: &FunctionArgs) -> Result<Value> {
+        self.exec_expr(args.args.get(1).expect("Missing arg 1"))
+            .await
+    }
+
+    async fn call_expr_third_arg(&mut self, args: &FunctionArgs) -> Result<Value> {
+        self.exec_expr(args.args.get(2).expect("Missing arg 2"))
+            .await
+    }
+
+    async fn call_expr_fourth_arg(&mut self, args: &FunctionArgs) -> Result<Value> {
+        self.exec_expr(args.args.get(3).expect("Missing arg 3"))
+            .await
+    }
+
+    async fn call_expr_fifth_arg(&mut self, args: &FunctionArgs) -> Result<Value> {
+        self.exec_expr(args.args.get(4).expect("Missing arg 4"))
+            .await
+    }
+
+    fn ensure_two_args(&self, args: &FunctionArgs, err: &str) -> Result<()> {
+        if args.len() != 2 {
+            return Err(DataFusionError::NotImplemented(err.into()));
+        }
+        Ok(())
+    }
+
+    fn ensure_three_args(&self, args: &FunctionArgs, err: &str) -> Result<()> {
+        if args.len() != 3 {
+            return Err(DataFusionError::NotImplemented(err.into()));
+        }
+        Ok(())
+    }
+
+    fn ensure_ge_three_args(&self, args: &FunctionArgs, err: &str) -> Result<()> {
+        if args.len() < 3 {
+            return Err(DataFusionError::NotImplemented(err.into()));
+        }
+        Ok(())
+    }
+
+    fn ensure_five_args(&self, args: &FunctionArgs, err: &str) -> Result<()> {
+        if args.len() != 5 {
+            return Err(DataFusionError::NotImplemented(err.into()));
+        }
+        Ok(())
+    }
+
+    fn parse_f64_else_err<T: Into<String>>(&self, value: &Value, err: T) -> Result<f64> {
+        match value {
+            Value::Float(f) => Ok(*f),
+            _ => Err(DataFusionError::NotImplemented(err.into())),
+        }
+    }
+    pub(super) async fn call_expr(
+        &mut self,
+        func: &Function,
+        args: &FunctionArgs,
+    ) -> Result<Value> {
+        use crate::functions::Func;
+
+        let func_name = Func::from_str(func.name).map_err(|_| {
+            DataFusionError::NotImplemented(format!("Unsupported function: {}", func.name))
+        })?;
+
+        // There are a few functions which need no arguments for e.g. time()
+        let functions_without_args: HashSet<&str> = HashSet::from_iter(vec![
+            "day_of_month",
+            "day_of_week",
+            "day_of_year",
+            "days_in_month",
+            "hour",
+            "minute",
+            "month",
+            "time",
+            "year",
+        ]);
+        let input = match functions_without_args.contains(func.name) {
+            true => match args.len() {
+                0 => {
+                    // Found no arg to pass to, lets use a `matrix(time())` as the arg.
+                    // https://prometheus.io/docs/prometheus/latest/querying/functions/#functions
+                    let timestamps = self.eval_ctx.timestamps();
+                    let samples: Vec<Sample> = timestamps
+                        .iter()
+                        .map(|&ts| Sample::new(ts, ts as f64))
+                        .collect();
+                    let default_now_matrix = vec![RangeValue {
+                        labels: Labels::default(),
+                        samples,
+                        exemplars: None,
+                        time_window: None,
+                    }];
+                    Value::Matrix(default_now_matrix)
+                }
+                1 => self.call_expr_first_arg(args).await?,
+
+                _ => {
+                    return Err(DataFusionError::NotImplemented(
+                        "Invalid args passed to the function".into(),
+                    ));
+                }
+            },
+            false => {
+                let last_arg = args
+                    .last()
+                    .expect("BUG: promql-parser should have validated function arguments");
+                self.exec_expr(&last_arg).await?
+            }
+        };
+
+        let start = std::time::Instant::now();
+        let result = match func_name {
+            Func::Abs => functions::abs(input)?,
+            Func::Absent => functions::absent(input, &self.eval_ctx)?,
+            Func::AbsentOverTime => functions::absent_over_time(input, &self.eval_ctx)?,
+            Func::AvgOverTime => functions::avg_over_time(input, &self.eval_ctx)?,
+            Func::Ceil => functions::ceil(input)?,
+            Func::Changes => functions::changes(input, &self.eval_ctx)?,
+            Func::Clamp => {
+                let err =
+                    "Invalid args, expected \"clamp(v instant-vector, min scalar, max scalar)\"";
+                self.ensure_three_args(args, err)?;
+
+                let input = self.call_expr_first_arg(args).await?;
+                let min = self.call_expr_second_arg(args).await?;
+                let max = self.call_expr_third_arg(args).await?;
+
+                let (min_f, max_f) = match (min, max) {
+                    (Value::Float(min), Value::Float(max)) => {
+                        if min > max {
+                            return Ok(Value::Matrix(vec![]));
+                        }
+                        (min, max)
+                    }
+                    _ => {
+                        return Err(DataFusionError::NotImplemented(err.into()));
+                    }
+                };
+                functions::clamp(input, min_f, max_f)?
+            }
+            Func::ClampMax => {
+                let err = "Invalid args, expected \"clamp(v instant-vector, max scalar)\"";
+                self.ensure_two_args(args, err)?;
+
+                let input = self.call_expr_first_arg(args).await?;
+                let max = self.call_expr_second_arg(args).await?;
+                let max_f = match max {
+                    Value::Float(max) => max,
+                    _ => {
+                        return Err(DataFusionError::NotImplemented(err.into()));
+                    }
+                };
+                functions::clamp(input, f64::MIN, max_f)?
+            }
+            Func::ClampMin => {
+                let err = "Invalid args, expected \"clamp(v instant-vector, min scalar)\"";
+                self.ensure_two_args(args, err)?;
+
+                let input = self.call_expr_first_arg(args).await?;
+                let min = self.call_expr_second_arg(args).await?;
+                let min_f = match min {
+                    Value::Float(min) => min,
+                    _ => {
+                        return Err(DataFusionError::NotImplemented(err.into()));
+                    }
+                };
+                functions::clamp(input, min_f, f64::MAX)?
+            }
+            Func::CountOverTime => functions::count_over_time(input, &self.eval_ctx)?,
+            Func::DayOfMonth => functions::day_of_month(input)?,
+            Func::DayOfWeek => functions::day_of_week(input)?,
+            Func::DayOfYear => functions::day_of_year(input)?,
+            Func::DaysInMonth => functions::days_in_month(input)?,
+            Func::Delta => functions::delta(input, &self.eval_ctx)?,
+            Func::Deriv => functions::deriv(input, &self.eval_ctx)?,
+            Func::Exp => functions::exp(input)?,
+            Func::Floor => functions::floor(input)?,
+            Func::HistogramCount => {
+                return Err(DataFusionError::NotImplemented(format!(
+                    "Unsupported Function: {func_name:?}"
+                )));
+            }
+            Func::HistogramFraction => {
+                return Err(DataFusionError::NotImplemented(format!(
+                    "Unsupported Function: {func_name:?}"
+                )));
+            }
+            Func::HistogramQuantile => {
+                let args = &args.args;
+                if args.len() != 2 {
+                    return Err(DataFusionError::Plan(format!(
+                        "{}: expected 2 arguments, got {}",
+                        func.name,
+                        args.len()
+                    )));
+                }
+                let phi = {
+                    match *args[0] {
+                        PromExpr::NumberLiteral(ref num) => num.val,
+                        _ => {
+                            return Err(DataFusionError::Plan(format!(
+                                "{}: the first argument must be a number",
+                                func.name
+                            )));
+                        }
+                    }
+                };
+
+                // Use range version if we have an eval context
+                functions::histogram_quantile(phi, input, &self.eval_ctx)?
+            }
+            Func::HistogramSum => {
+                return Err(DataFusionError::NotImplemented(format!(
+                    "Unsupported Function: {func_name:?}"
+                )));
+            }
+            Func::HoltWinters => {
+                let err =
+                    "Invalid args, expected \"holt_winters(v range-vector, sf scalar, tf scalar)\"";
+                self.ensure_three_args(args, err)?;
+
+                let input = self.call_expr_first_arg(args).await?;
+                let sf = self.call_expr_second_arg(args).await?;
+                let tf = self.call_expr_third_arg(args).await?;
+
+                let scaling_factor = self.parse_f64_else_err(&sf, err)?;
+                let trend_factor = self.parse_f64_else_err(&tf, err)?;
+
+                functions::holt_winters(input, scaling_factor, trend_factor, &self.eval_ctx)?
+            }
+            Func::Hour => functions::hour(input)?,
+            Func::Idelta => functions::idelta(input, &self.eval_ctx)?,
+            Func::Increase => functions::increase(input, &self.eval_ctx)?,
+            Func::Irate => functions::irate(input, &self.eval_ctx)?,
+            Func::LabelJoin => {
+                let err = "Invalid args, expected \"label_join(v instant-vector, dst string, sep string, src_1 string, src_2 string, ...)\"";
+                self.ensure_ge_three_args(args, err)?;
+
+                let input = self.call_expr_first_arg(args).await?;
+                let dst_label = self.call_expr_second_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid destination label found".into()),
+                )?;
+                let separator = self.call_expr_third_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid separator label found".into()),
+                )?;
+
+                let mut source_labels = vec![];
+                for each_src in args.args[3..].iter() {
+                    if let Value::String(label) = self.exec_expr(each_src).await.unwrap() {
+                        source_labels.push(label);
+                    };
+                }
+                if source_labels.is_empty() {
+                    return Err(DataFusionError::NotImplemented(
+                        "source labels can not be empty or invalid".into(),
+                    ));
+                }
+                functions::label_join(input, &dst_label, &separator, source_labels)?
+            }
+            Func::LabelReplace => {
+                let err = "Invalid args, expected \"label_replace(v instant-vector, dst_label string, replacement string, src_label string, regex string)\"";
+
+                self.ensure_five_args(args, err)?;
+                let input = self.call_expr_first_arg(args).await?;
+
+                let dst_label = self.call_expr_second_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid destination label found".into()),
+                )?;
+                let replacement = self.call_expr_third_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid replacement string found".into()),
+                )?;
+
+                let src_label = self.call_expr_fourth_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid source label string found".into()),
+                )?;
+
+                let regex = self.call_expr_fifth_arg(args).await?.get_string().ok_or(
+                    DataFusionError::NotImplemented("Invalid regex string found".into()),
+                )?;
+
+                functions::label_replace(input, &dst_label, &replacement, &src_label, &regex)?
+            }
+            Func::LastOverTime => functions::last_over_time(input, &self.eval_ctx)?,
+            Func::Ln => functions::ln(input)?,
+            Func::Log10 => functions::log10(input)?,
+            Func::Log2 => functions::log2(input)?,
+            Func::MaxOverTime => functions::max_over_time(input, &self.eval_ctx)?,
+            Func::MinOverTime => functions::min_over_time(input, &self.eval_ctx)?,
+            Func::Minute => functions::minute(input)?,
+            Func::Month => functions::month(input)?,
+            Func::PredictLinear => {
+                let err = "Invalid args, expected \"predict_linear(v range-vector, t scalar)\"";
+
+                self.ensure_two_args(args, err)?;
+                let input = self.call_expr_first_arg(args).await?;
+
+                let prediction_steps = self.call_expr_second_arg(args).await?.get_float().ok_or(
+                    DataFusionError::NotImplemented(
+                        "Invalid prediction_steps, f64 expected".into(),
+                    ),
+                )?;
+                functions::predict_linear(input, prediction_steps, &self.eval_ctx)?
+            }
+            Func::QuantileOverTime => {
+                let err = "Invalid args, expected \"quantile_over_time(scalar, range-vector)\"";
+
+                self.ensure_two_args(args, err)?;
+                let phi_quantile = match self.call_expr_first_arg(args).await {
+                    Ok(Value::Float(v)) => v,
+                    _ => {
+                        return Err(DataFusionError::Plan(
+                            "[quantile] param must be a NumberLiteral".into(),
+                        ));
+                    }
+                };
+                let input = self.call_expr_second_arg(args).await?;
+                functions::quantile_over_time(phi_quantile, input, &self.eval_ctx)?
+            }
+            Func::Rate => functions::rate(input, &self.eval_ctx)?,
+            Func::Resets => functions::resets(input, &self.eval_ctx)?,
+            Func::Round => functions::round(input)?,
+            Func::Scalar => functions::scalar(input, &self.eval_ctx)?,
+            Func::Sgn => functions::sgn(input)?,
+            Func::Sort => {
+                return Err(DataFusionError::NotImplemented(format!(
+                    "Unsupported Function: {func_name:?}"
+                )));
+            }
+            Func::SortDesc => {
+                return Err(DataFusionError::NotImplemented(format!(
+                    "Unsupported Function: {func_name:?}"
+                )));
+            }
+            Func::Sqrt => functions::sqrt(input)?,
+            Func::StddevOverTime => functions::stddev_over_time(input, &self.eval_ctx)?,
+            Func::StdvarOverTime => functions::stdvar_over_time(input, &self.eval_ctx)?,
+            Func::SumOverTime => functions::sum_over_time(input, &self.eval_ctx)?,
+            // TODO: check this implementation
+            Func::Time => Value::Float((self.eval_ctx.start / 1_000_000) as f64),
+            Func::Timestamp => functions::timestamp(input)?,
+            Func::Vector => functions::vector(input, &self.eval_ctx)?,
+            Func::Year => functions::year(input)?,
+        };
+        log::info!(
+            "[trace_id: {}] [PromQL Timing] call_expr({}) execution took: {:?}",
+            self.trace_id,
+            func.name,
+            start.elapsed()
+        );
+        Ok(result)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use promql_parser::parser::{FunctionArgs, NumberLiteral};
+
+    use super::*;
+    use crate::{engine::test_util::*, exec::PromqlContext};
+
+    #[test]
+    fn test_ensure_two_args() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 1.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 2.0 })),
+            ],
+        };
+        let result = engine.ensure_two_args(&args, "test error");
+        assert!(result.is_ok());
+
+        let args = FunctionArgs {
+            args: vec![Box::new(PromExpr::NumberLiteral(NumberLiteral {
+                val: 1.0,
+            }))],
+        };
+        let result = engine.ensure_two_args(&args, "test error");
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_ensure_three_args() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 1.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 2.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 3.0 })),
+            ],
+        };
+        let result = engine.ensure_three_args(&args, "test error");
+        assert!(result.is_ok());
+
+        let args = FunctionArgs {
+            args: vec![Box::new(PromExpr::NumberLiteral(NumberLiteral {
+                val: 1.0,
+            }))],
+        };
+        let result = engine.ensure_three_args(&args, "test error");
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_ensure_ge_three_args() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 1.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 2.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 3.0 })),
+            ],
+        };
+        let result = engine.ensure_ge_three_args(&args, "test error");
+        assert!(result.is_ok());
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 1.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 2.0 })),
+            ],
+        };
+        let result = engine.ensure_ge_three_args(&args, "test error");
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_ensure_five_args() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 1.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 2.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 3.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 4.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 5.0 })),
+            ],
+        };
+        let result = engine.ensure_five_args(&args, "test error");
+        assert!(result.is_ok());
+
+        let args = FunctionArgs {
+            args: vec![Box::new(PromExpr::NumberLiteral(NumberLiteral {
+                val: 1.0,
+            }))],
+        };
+        let result = engine.ensure_five_args(&args, "test error");
+        assert!(result.is_err());
+    }
+
+    #[test]
+    fn test_parse_f64_else_err() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let value = Value::Float(42.0);
+        let result = engine.parse_f64_else_err(&value, "test error");
+        assert!(result.is_ok());
+        assert_eq!(result.unwrap(), 42.0);
+
+        let value = Value::String("not a float".to_string());
+        let result = engine.parse_f64_else_err(&value, "test error");
+        assert!(result.is_err());
+    }
+
+    #[tokio::test]
+    async fn test_call_expr_first_arg() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let mut engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![Box::new(PromExpr::NumberLiteral(NumberLiteral {
+                val: 42.0,
+            }))],
+        };
+        let result = engine.call_expr_first_arg(&args).await;
+        assert!(result.is_ok());
+
+        if let Ok(Value::Float(val)) = result {
+            assert_eq!(val, 42.0);
+        } else {
+            panic!("Expected Value::Float");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_call_expr_second_arg() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let mut engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 10.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 42.0 })),
+            ],
+        };
+        let result = engine.call_expr_second_arg(&args).await;
+        assert!(result.is_ok());
+
+        if let Ok(Value::Float(val)) = result {
+            assert_eq!(val, 42.0);
+        } else {
+            panic!("Expected Value::Float");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_call_expr_third_arg() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let mut engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 10.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 20.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 42.0 })),
+            ],
+        };
+        let result = engine.call_expr_third_arg(&args).await;
+        assert!(result.is_ok());
+
+        if let Ok(Value::Float(val)) = result {
+            assert_eq!(val, 42.0);
+        } else {
+            panic!("Expected Value::Float");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_call_expr_fourth_arg() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let mut engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 10.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 20.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 30.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 42.0 })),
+            ],
+        };
+        let result = engine.call_expr_fourth_arg(&args).await;
+        assert!(result.is_ok());
+
+        if let Ok(Value::Float(val)) = result {
+            assert_eq!(val, 42.0);
+        } else {
+            panic!("Expected Value::Float");
+        }
+    }
+
+    #[tokio::test]
+    async fn test_call_expr_fifth_arg() {
+        let trace_id = "test_trace";
+        let org_id = "test_org";
+        let mut engine = Engine::new(
+            trace_id,
+            Arc::new(PromqlContext::new(
+                create_test_query_ctx(trace_id, org_id, 30),
+                SimpleMockProvider,
+                vec![],
+            )),
+            create_test_eval_ctx(),
+        );
+
+        let args = FunctionArgs {
+            args: vec![
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 10.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 20.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 30.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 40.0 })),
+                Box::new(PromExpr::NumberLiteral(NumberLiteral { val: 42.0 })),
+            ],
+        };
+        let result = engine.call_expr_fifth_arg(&args).await;
+        assert!(result.is_ok());
+
+        if let Ok(Value::Float(val)) = result {
+            assert_eq!(val, 42.0);
+        } else {
+            panic!("Expected Value::Float");
+        }
+    }
+}
